Workaround for NPM download module corrupting $HOME environment variable
